Extract the AstSmell class
