+++ good_smell/__init__.py
@@ -1,5 +1,6 @@
 from .smell_warning import SmellWarning
 from .lint_smell import LintSmell
+from .ast_smell import AstSmell
 from .smells import implemented_smells
 from . import smells # Allow importing good_smell.smells
 from .flake8_ext import GoodSmellFlake8
